Added a background class
I've added a class called backgrounds. it is going to describe how the
backgrounds are going to behave with movement and wich backgrounds to
draw out.

I've added 3 methods:
The draw method is going to draw out the current backgrounds

The move method is going to mvoe the backgrounds down on the screen

The change_current_bg is going to change the backgrounds that is drawn
out to make it easier to run with as few backgrounds activated at the
same time