Added code for save file
added a set and get save, i currenctly only write the latest highscore
into the save(the get save is not used and does not work). Next step is
to make it save the highest score and then make it a top ten list. After
that we can put it in so you can see the highscores in game.