fixed animation bug and redid timer
I've fixed the bug where the animation wont go. This was fixed by
resetting the update_round to 0 when a minute passes. When the game
restarts the update_round also resets.

I needed to remake the way that the time is calculated to a better more
clean way. When seconds go up to 59, add a minute and resets the seconds